+++ POS(CapstoneProject)/wwwroot/js/Cashier/Sales.js
@@ -282,13 +282,53 @@ document.querySelector('#inputDiscount').addEventListener('change', function () 
 
 $('.cancelBtn').click(function () {
 
-    CheckOutToggle();
 
-    checkOutList.splice(0);
+    if (checkOutList.length == 0) {
 
-    DisplayCheckOut();
-    cartCount();
-    localStorage.setItem('checkoutListCashier', JSON.stringify(checkOutList));
+        popUpMessageSales("Check Out list is empty", "error");
+    }
+    else {
+        Swal.fire({
+            icon: "question",
+            title: "Are you sure you want to cancel? <br>",
+
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes",
+            customClass: {
+                icon: 'custom-icon',
+                title: 'swal-salesCancel-title',
+
+            }
+
+        }).then((result) => {
+
+            if (result.isConfirmed) {
+                const Toast = Swal.mixin({
+                    toast: true,
+                    position: "top-end",
+                    width: 300,
+                    showConfirmButton: false,
+                    timer: 1500,
+
+                });
+                Toast.fire({
+                    icon: "success",
+                    title: "Check out Canceled"
+                });
+
+                checkOutList.splice(0);
+
+                DisplayCheckOut();
+                cartCount();
+                localStorage.setItem('checkoutListCashier', JSON.stringify(checkOutList));
+            }
+        });
+    }
+    /*   CheckOutToggle();*/
+   
+
 
 })
 
@@ -296,7 +336,7 @@ $('.cancelBtn').click(function () {
 $('.payBtn').click(function () {
 
     if (checkOutList.length === 0) {
-        popUpMessageSales("Your check out list is empty", "error");
+        popUpMessageSales("Check Out list is empty", "error");
 
     } else {
         $('#paymentModal').modal('toggle')
